+++ mock_drivers/mock_lightbar_driver/include/mock_lightbar_driver/MockLightBarDriver.hpp
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2019-2020 LEIDOS.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+
+
 #include <ros/ros.h>
 #include <ros/node_handle.h>
 #include <iostream>
@@ -20,9 +38,10 @@
 #include <cav_srvs/SetLightsRequest.h>
 #include <cav_srvs/SetLightsResponse.h>
 #include <cav_msgs/LightBarStatus.h>
+#include "carma_utils/CARMANodeHandle.h"
 using namespace std;
 
-namespace translated_mock_driver_testing{
+namespace mock_drivers{
     class MockLightBarDriver  {
         
 
@@ -84,7 +103,7 @@ namespace translated_mock_driver_testing{
             vector<string> getDriverApi();
 
           public:
-             MockLightBarDriver(ros::NodeHandle node);
+             MockLightBarDriver(ros::CARMANodeHandle node);
 
             //publishes lightbar status
             void publishData();
