+++ mock_drivers/mock_lightbar_driver/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>translated_mock_driver_testing</name>
-  <description>translated_mock_driver_testing package</description>
+  <name>mock_lightbar_driver</name>
+  <description>The mock_lightbar_driver package</description>
   <version>1.0.0</version>
   <buildtool_depend>catkin</buildtool_depend>
   <license>APACHE 2.0</license>
@@ -9,6 +9,8 @@
   <depend>roscpp</depend>
   <depend>cav_srvs</depend>
   <depend>cav_msgs</depend>
+  <depend>carma_utils</depend>
+  
 
 
 </package>
