+++ mock_drivers/mock_lightbar_driver/test/test_mock_light_bar_driver.cpp
@@ -1,15 +1,29 @@
+/*
+ * Copyright (C) 2019-2020 LEIDOS.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
 #include <gtest/gtest.h>
-#include "translated_mock_driver_testing/MockLightBarDriver.hpp"
+#include "mock_lightbar_driver/MockLightBarDriver.hpp"
 #include <ros/ros.h>
 #include <ros/service_server.h>
-using namespace translated_mock_driver_testing;
+#include "carma_utils/CARMANodeHandle.h"
+using namespace mock_drivers;
 
 
-
-    
-
-    TEST(GetLightsBasicTest,test){
-        ros::NodeHandle n = ros::NodeHandle();
+    TEST(GetLightsBasicTest,testGetLights){
+        ros::CARMANodeHandle n;
         ros::ServiceClient client = n.serviceClient<cav_srvs::GetLights>("lightbar/get_lights");
         cav_srvs::GetLights light;
         ROS_INFO_STREAM(" " << client.getService());
@@ -23,8 +37,8 @@ using namespace translated_mock_driver_testing;
          }
     }
 
-    TEST(SetLightsTest,test){
-         ros::NodeHandle n = ros::NodeHandle();
+    TEST(SetLightsTest,testSetLights){
+         ros::CARMANodeHandle n;
         ros::ServiceClient client = n.serviceClient<cav_srvs::SetLights>("lightbar/set_lights");
         ros::ServiceClient client2 = n.serviceClient<cav_srvs::GetLights>("lightbar/get_lights");
         cav_srvs::GetLights light1;
@@ -46,8 +60,8 @@ using namespace translated_mock_driver_testing;
         } 
     }
 
-    TEST(GetDriverAPITest,test){
-        ros::NodeHandle n = ros::NodeHandle();
+    TEST(GetDriverAPITest,testGetDriverAPI){
+        ros::CARMANodeHandle n;
         ros::ServiceClient client = n.serviceClient<cav_srvs::GetDriverApi>("get_driver_api");
         cav_srvs::GetDriverApi api;
         if (client.call(api)){
@@ -56,8 +70,8 @@ using namespace translated_mock_driver_testing;
         }
     }
 
-    TEST(GetDriverStatusTest,test){
-        ros::NodeHandle n = ros::NodeHandle();
+    TEST(GetDriverStatusTest,testGetDriverStatus){
+        ros::CARMANodeHandle n;
         ros::ServiceClient client = n.serviceClient<cav_srvs::GetDriverStatus>("get_status");
         cav_srvs::GetDriverStatus status;
         if (client.call(status)){
@@ -75,8 +89,8 @@ using namespace translated_mock_driver_testing;
 
     }
 
-    TEST(DriverStatusPubTest,test){
-        ros::NodeHandle n = ros::NodeHandle();
+    TEST(DriverStatusPubTest,testDriverStatusPublish){
+        ros::CARMANodeHandle n;
         cav_msgs::DriverStatus status;
         ros::Subscriber sub = n.subscribe("driver_discovery",10,&discovery_cb);
         ros::spinOnce();
@@ -87,8 +101,8 @@ using namespace translated_mock_driver_testing;
                 && status.right_arrow == 1 && status.sides_solid == 0 && status.green_solid == 0
                 && status.yellow_solid == 0 && status.takedown == 0);
     }
-    TEST(LightBarStatusPubTest, test){
-        ros::NodeHandle n = ros::NodeHandle();
+    TEST(LightBarStatusPubTest, testLightBarStatusPublish){
+        ros::CARMANodeHandle n;
         cav_msgs::LightBarStatus status;
         ros::Subscriber sub = n.subscribe("lightbar/light_bar_status",10,&lightbar_cb);
         ros::spinOnce();
@@ -97,12 +111,8 @@ using namespace translated_mock_driver_testing;
 
     int main(int argc, char** argv){
         testing::InitGoogleTest(&argc,argv);
-        ros::init(argc,argv, "MockLightBarDriver_Test");
-        //ros::NodeHandle f = ros::NodeHandle();
-        //MockLightBarDriver g(f);
+        ros::init(argc,argv, "test_mock_light_bar_driver");
         auto result = RUN_ALL_TESTS();
-       // ros::spin();
-       // ros::shutdown();
         return result;
 
     }
